Update README.md
If you wish to forego installing ganache and redis, you can enable a simulator mode. What this does is bypass the eth_worker and any queued jobs, and instead returns dummy responses to any functions relying on eth_worker. This _will_ make your database fall out of sync with any ganache instance you have set up, so use this with care, but it is very useful in eliminating dependencies when working on any features in the API or frontend